requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
black==23.3.0
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
decorator==5.1.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
dulwich==0.21.5
executing==1.2.0
filelock==3.12.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
installer==0.7.0
ipdb==0.13.13
ipython==8.14.0
jaraco.classes==3.2.3
jedi==0.18.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
matplotlib-inline==0.1.6
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.6
platformdirs==3.8.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prompt-toolkit==3.0.38
psycopg2-binary-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomlkit==0.11.8
traitlets==5.9.0
trove-classifiers==2023.5.24
typing_extensions==4.7.1
tzdata==2023.3
upgrade-requirements==1.7.0
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
zipp==3.15.0
whitenoise==6.5.0
